fix(e2e): consolidate test suites to reduce execution time
